+++ time-series/time-series-api/src/test/java/com/powsybl/timeseries/StoredDoubleTimeSeriesTest.java
@@ -25,12 +25,7 @@ import java.util.Collections;
 import java.util.List;
 
 import static java.lang.Double.NaN;
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertInstanceOf;
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * @author Geoffroy Jamgotchian {@literal <geoffroy.jamgotchian at rte-france.com>}
@@ -413,7 +408,7 @@ class StoredDoubleTimeSeriesTest {
     }
 
     @Test
-    void toArrayWhenTimeSeriesData() {
+    void testToArray() {
         // Given
         TimeSeriesIndex index = Mockito.mock(TimeSeriesIndex.class);
         Mockito.when(index.getPointCount()).thenReturn(8);
@@ -432,17 +427,10 @@ class StoredDoubleTimeSeriesTest {
         assertArrayEquals(new double[]{1d, 2d, 3d, 4d, 5d, 6d, 7d, 8d}, timeSeriesArray, 0d);
         assertArrayEquals(new double[]{1d, 2d, 3d, 4d, NaN, NaN, NaN, NaN}, tsArray1, 0d);
         assertArrayEquals(new double[]{NaN, NaN, NaN, NaN, 5d, 6d, 7d, 8d}, tsArray2, 0d);
-
-        double originalAt3 = timeSeries.get(3);
-        assertEquals(4.0, originalAt3, 0d);
-
-        double splitAt3 = chunks.get(0).get(3);
-        assertEquals(4.0, splitAt3, 0d);
-        assertEquals(originalAt3, splitAt3, 0d);
     }
 
     @Test
-    void toCompactArrayShouldKeepOriginalIndexeWithoutNaN() {
+    void testToCompactArray() {
         // Given
         TimeSeriesIndex index = Mockito.mock(TimeSeriesIndex.class);
         Mockito.when(index.getPointCount()).thenReturn(8);
@@ -461,17 +449,10 @@ class StoredDoubleTimeSeriesTest {
         assertArrayEquals(new double[]{1d, 2d, 3d, 4d, 5d, 6d, 7d, 8d}, timeSeriesCompactArray, 0d);
         assertArrayEquals(new double[]{1d, 2d, 3d, 4d}, tsCompactArray1, 0d);
         assertArrayEquals(new double[]{5d, 6d, 7d, 8d}, tsCompactArray2, 0d);
-
-        double originalAt3 = timeSeries.get(3);
-        assertEquals(4.0, originalAt3, 0d);
-
-        double splitAt3 = chunks.get(0).get(3);
-        assertEquals(4.0, splitAt3, 0d);
-        assertEquals(originalAt3, splitAt3, 0d);
     }
 
     @Test
-    void toCompactArrayWhenNaNExistsAtTheMiddle() {
+    void testToCompactArrayWhenNaNExistsAtTheMiddle() {
         TimeSeriesIndex index = Mockito.mock(TimeSeriesIndex.class);
         Mockito.when(index.getPointCount()).thenReturn(9);
         TimeSeriesMetadata metadata = new TimeSeriesMetadata("ts1", TimeSeriesDataType.DOUBLE, index);
