+++ time-series/time-series-api/src/test/java/com/powsybl/timeseries/StringTimeSeriesTest.java
@@ -173,7 +173,7 @@ class StringTimeSeriesTest {
     }
 
     @Test
-    void toArrayWhenTimeSeriesData() {
+    void testToArray() {
         // Given
         TimeSeriesIndex index = Mockito.mock(TimeSeriesIndex.class);
         Mockito.when(index.getPointCount()).thenReturn(8);
@@ -192,17 +192,10 @@ class StringTimeSeriesTest {
         assertArrayEquals(new String[]{"a", "b", "c", "d", "e", "f", "g", "h"}, timeSeriesArray);
         assertArrayEquals(new String[]{"a", "b", "c", "d", null, null, null, null}, tsArray1);
         assertArrayEquals(new String[]{null, null, null, null, "e", "f", "g", "h"}, tsArray2);
-
-        String originalAt3 = timeSeries.get(3);
-        assertEquals("d", originalAt3);
-
-        String splitAt3 = chunks.get(0).get(3);
-        assertEquals("d", splitAt3);
-        assertEquals(originalAt3, splitAt3);
     }
 
     @Test
-    void toCompactArrayShouldKeepOriginalIndexeWithoutNull() {
+    void testToCompactArray() {
         // Given
         TimeSeriesIndex index = Mockito.mock(TimeSeriesIndex.class);
         Mockito.when(index.getPointCount()).thenReturn(8);
@@ -221,17 +214,10 @@ class StringTimeSeriesTest {
         assertArrayEquals(new String[]{"a", "b", "c", "d", "e", "f", "g", "h"}, timeSeriesArray);
         assertArrayEquals(new String[]{"a", "b", "c", "d"}, tsArray1);
         assertArrayEquals(new String[]{"e", "f", "g", "h"}, tsArray2);
-
-        String originalAt3 = timeSeries.get(3);
-        assertEquals("d", originalAt3);
-
-        String splitAt3 = chunks.get(0).get(3);
-        assertEquals("d", splitAt3);
-        assertEquals(originalAt3, splitAt3);
     }
 
     @Test
-    void toCompactArrayWhenNaNExistsAtTheMiddle() {
+    void testToCompactArrayWhenNaNExistsAtTheMiddle() {
         TimeSeriesIndex index = Mockito.mock(TimeSeriesIndex.class);
         Mockito.when(index.getPointCount()).thenReturn(9);
         TimeSeriesMetadata metadata = new TimeSeriesMetadata("ts1", TimeSeriesDataType.DOUBLE, index);
@@ -240,6 +226,30 @@ class StringTimeSeriesTest {
         StringTimeSeries timeSeries = new StringTimeSeries(metadata, chunk1, chunk2);
         assertArrayEquals(new String[] {"a", "b", "c", null, null, null, "g", "h", null}, timeSeries.toArray());
         assertArrayEquals(new String[] {"a", "b", "c", null, null, null, "g", "h"}, timeSeries.toCompactArray());
+    }
+
+    @Test
+    void testGetByIndex() {
+        RegularTimeSeriesIndex index = RegularTimeSeriesIndex.create(Interval.parse("2015-01-01T00:00:00Z/2015-01-01T01:45:00Z"), Duration.ofMinutes(15));
+        TimeSeriesMetadata metadata = new TimeSeriesMetadata("ts1", TimeSeriesDataType.DOUBLE, index);
+        UncompressedStringDataChunk chunk1 = new UncompressedStringDataChunk(0, new String[]{"a", "b", "c", "d", "e", "f"});
+        UncompressedStringDataChunk chunk2 = new UncompressedStringDataChunk(6, new String[]{"g", "h"});
+        StringTimeSeries timeSeries = new StringTimeSeries(metadata, chunk1, chunk2);
+        List<StringTimeSeries> splitTimeSeries = timeSeries.split(4);
+        StringTimeSeries timeSeries1 = splitTimeSeries.get(0);
+        StringTimeSeries timeSeries2 = splitTimeSeries.get(1);
+        // Original time series
+        assertTimeSerie(timeSeries, "b", "f", "g");
+        // First split time series
+        assertTimeSerie(timeSeries1, "b", null, null);
+        // Second split time series
+        assertTimeSerie(timeSeries2, null, "f", "g");
+    }
+
+    private static void assertTimeSerie(StringTimeSeries timeSeries, String expectedAtIndex1, String expectedAtIndex5, String expectedAtIndex6) {
+        assertEquals(expectedAtIndex1, timeSeries.get(1));
+        assertEquals(expectedAtIndex5, timeSeries.get(5));
+        assertEquals(expectedAtIndex6, timeSeries.get(6));
     }
 
 }
